frontend: Replace home with /demos/new_home
